🚧 Add WebMessageChannel transport for AWC.
