+++ lib/infrastructure/rpc/websocket_server.dart
@@ -2,100 +2,9 @@ import 'dart:async';
 import 'dart:developer';
 import 'dart:io';
 
-import 'package:aewallet/domain/rpc/subscription.dart';
-import 'package:aewallet/infrastructure/rpc/add_service/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/dto/rpc_command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/dto/rpc_request.dart';
-import 'package:aewallet/infrastructure/rpc/dto/rpc_subscription.dart';
-import 'package:aewallet/infrastructure/rpc/get_accounts/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/get_current_account/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/get_endpoint/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/get_services_from_keychain/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/keychain_derive_address/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/keychain_derive_keypair/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/refresh_current_account/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/send_transaction/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/sign_transactions/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/sub_account/command_handler.dart';
-import 'package:aewallet/infrastructure/rpc/sub_current_account/command_handler.dart';
+import 'package:aewallet/infrastructure/rpc/awc_json_rpc_server.dart';
 import 'package:flutter/foundation.dart';
-import 'package:json_rpc_2/json_rpc_2.dart';
 import 'package:web_socket_channel/io.dart';
-
-/// A [Peer] composition which handles subscription requests
-class _SubscribablePeer {
-  _SubscribablePeer(this._peer);
-
-  final Peer _peer;
-
-  final Map<String, StreamSubscription> _subscriptions = {};
-
-  Future<void> close() => _peer.close();
-
-  void registerMethod(String name, Function callback) =>
-      _peer.registerMethod(name, callback);
-
-  void registerSubscriptionMethod(
-    String name,
-    Future<RPCSubscriptionDTO> Function(dynamic params) subscriptionCallback,
-  ) {
-    registerMethod(
-      name,
-      (params) async {
-        final result = await subscriptionCallback(params);
-
-        final streamSubscription = result.updates.listen((value) {
-          _peer.sendNotification('addSubscriptionNotification', value);
-        });
-        _registerSubscription(result.id, streamSubscription);
-
-        return result.toJson();
-      },
-    );
-  }
-
-  void registerUnsubscriptionMethod(
-    String name,
-  ) {
-    registerMethod(
-      name,
-      (params) async {
-        final requestDTO = RPCRequestDTO.fromJson(
-          params.value,
-        );
-
-        final unsubscribeCommand = RPCUnsubscribeCommandDTO.fromJson(
-          requestDTO.payload,
-        );
-        _removeSubscription(unsubscribeCommand.subscriptionId);
-        return {};
-      },
-    );
-  }
-
-  void registerFallback(Function(Parameters parameters) callback) =>
-      _peer.registerFallback(callback);
-
-  void _removeSubscription(String subscriptionId) {
-    _subscriptions.remove(subscriptionId);
-  }
-
-  void _registerSubscription(String subscriptionId, StreamSubscription sub) {
-    _subscriptions[subscriptionId] = sub;
-  }
-
-  void _cleanupSubscriptions() {
-    for (final subscription in _subscriptions.entries) {
-      subscription.value.cancel();
-    }
-    _subscriptions.clear();
-  }
-
-  Future listen() async {
-    await _peer.listen();
-    _cleanupSubscriptions();
-  }
-}
 
 class ArchethicWebsocketRPCServer {
   ArchethicWebsocketRPCServer();
@@ -110,7 +19,7 @@ class ArchethicWebsocketRPCServer {
   }
 
   HttpServer? _runningHttpServer;
-  final Set<_SubscribablePeer> _openedSockets = {};
+  final Set<AWCJsonRPCServer> _openedSockets = {};
 
   bool get isRunning => _openedSockets.isNotEmpty || _runningHttpServer != null;
 
@@ -135,71 +44,7 @@ class ArchethicWebsocketRPCServer {
           final socket = await WebSocketTransformer.upgrade(request);
           final channel = IOWebSocketChannel(socket);
 
-          final peerServer = _SubscribablePeer(Peer(channel.cast<String>()))
-            ..registerMethod(
-              'sendTransaction',
-              (params) => _handle(RPCSendTransactionCommandHandler(), params),
-            )
-            ..registerMethod(
-              'getEndpoint',
-              (params) => _handle(RPCGetEndpointCommandHandler(), params),
-            )
-            ..registerMethod(
-              'refreshCurrentAccount',
-              (params) =>
-                  _handle(RPCRefreshCurrentAccountCommandHandler(), params),
-            )
-            ..registerMethod(
-              'getCurrentAccount',
-              (params) => _handle(RPCGetCurrentAccountCommandHandler(), params),
-            )
-            ..registerMethod(
-              'getAccounts',
-              (params) => _handle(RPCGetAccountsCommandHandler(), params),
-            )
-            ..registerSubscriptionMethod(
-              'subscribeAccount',
-              (params) => _handleSubscription(
-                RPCSubscribeAccountCommandHandler(),
-                params,
-              ),
-            )
-            ..registerUnsubscriptionMethod(
-              'unsubscribeAccount',
-            )
-            ..registerSubscriptionMethod(
-              'subscribeCurrentAccount',
-              (params) => _handleSubscription(
-                RPCSubscribeCurrentAccountCommandHandler(),
-                params,
-              ),
-            )
-            ..registerUnsubscriptionMethod(
-              'unsubscribeCurrentAccount',
-            )
-            ..registerMethod(
-              'addService',
-              (params) => _handle(RPCAddServiceCommandHandler(), params),
-            )
-            ..registerMethod(
-              'getServicesFromKeychain',
-              (params) =>
-                  _handle(RPCGetServicesFromKeychainCommandHandler(), params),
-            )
-            ..registerMethod(
-              'keychainDeriveKeypair',
-              (params) =>
-                  _handle(RPCKeychainDeriveKeypairCommandHandler(), params),
-            )
-            ..registerMethod(
-              'keychainDeriveAddress',
-              (params) =>
-                  _handle(RPCKeychainDeriveAddressCommandHandler(), params),
-            )
-            ..registerMethod(
-              'signTransactions',
-              (params) => _handle(RPCSignTransactionsCommandHandler(), params),
-            );
+          final peerServer = AWCJsonRPCServer(channel.cast<String>());
 
           _openedSockets.add(peerServer);
           await peerServer.listen();
@@ -242,64 +87,6 @@ class ArchethicWebsocketRPCServer {
           error: error,
           stackTrace: stack,
           name: logName,
-        );
-      },
-    );
-  }
-
-  Future<Map<String, dynamic>> _handle(
-    RPCCommandHandler commandHandler,
-    Parameters params,
-  ) async {
-    final result = await commandHandler.handle(params.value);
-    return result.map(
-      success: commandHandler.resultFromModel,
-      failure: (failure) {
-        log(
-          'Command failed',
-          name: logName,
-          error: failure,
-        );
-
-        throw RpcException(
-          failure.code,
-          failure.message ?? 'Command failed',
-        );
-      },
-    );
-  }
-
-  Future<RPCSubscriptionDTO> _handleSubscription(
-    RPCSubscriptionHandler commandHandler,
-    Parameters params,
-  ) async {
-    final result = await commandHandler.handle(params.value);
-    return result.map(
-      success: (success) {
-        success as RPCSubscription;
-        return RPCSubscriptionDTO(
-          id: success.id,
-          updates: success.updates
-              .map(
-                (update) => RPCSubscriptionUpdateDTO(
-                  subscriptionId: success.id,
-                  data: commandHandler.notificationFromModel(update),
-                ),
-              )
-              .distinct()
-              .map((dto) => dto.toJson()),
-        );
-      },
-      failure: (failure) {
-        log(
-          'Command failed',
-          name: logName,
-          error: failure,
-        );
-
-        throw RpcException(
-          failure.code,
-          failure.message ?? 'Command failed',
         );
       },
     );
