+++ pubspec.yaml
@@ -113,6 +113,9 @@ dependencies:
   # A flutter implementation of React hooks. It adds a new kind of widget with enhanced code reuse
   flutter_hooks: ^0.20.0
 
+  # A WebView
+  flutter_inappwebview: ^6.0.0
+
   # Flutter jailbreak and root detection plugin. This plugin wraps Rootbeer for use on Android and DTTJailbreakDetection for use on iOS.
   flutter_jailbreak_detection: ^1.10.0
 
@@ -235,6 +238,8 @@ dependencies:
 
   # Dartlang port of socket.io-client for web, flutter, dartvm to use
   socket_io_client: ^2.0.3+1
+
+  stream_channel: any
 
   # Lock mechanism to prevent concurrent access to asynchronous code.
   synchronized: ^3.1.0
